# docs/swagger.yaml
definitions:
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.ListParams:
    properties:
      page:
        type: integer
      size:
        type: integer
    type: object
...
  model.UserDTOAdd:
    properties:
      account:
        type: string
      password:
        type: string
      re_password:
        type: string
    required:
    - account
    - password
    - re_password
    type: object
  model.UserDTOLogin:
    properties:
      account:
        type: string
      password:
        type: string
    required:
    - account
    - password
    type: object
  model.UserDTORegister:
    properties:
      account:
        type: string
      password:
        type: string
      re_password:
        type: string
    required:
    - account
    - password
    - re_password
    type: object
  model.UserDTOUpdateByAdmin:
    properties:
      account:
        type: string
      description:
        type: string
      email:
        type: string
      gender:
        type: boolean
      password:
        type: string
      phone:
        type: string
      re_password:
        type: string
      user_id:
        example: "0"
        type: string
      user_role:
        type: integer
    required:
    - account
    - user_id
    type: object
  model.UserDTOUpdateBySelf:
    properties:
      account:
        type: string
      description:
        type: string
      email:
        type: string
      gender:
        type: boolean
      password:
        type: string
      phone:
        type: string
      re_password:
        type: string
      user_id:
        example: "0"
        type: string
    required:
    - account
    - user_id
    type: object
info:
  contact: {}
paths:
...
  /user/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新用户的数据
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.UserDTOAdd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员添加用户
      tags:
      - 用户相关接口
  /user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要删除的user_id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员根据user_id删除用户
      tags:
      - 用户相关接口
  /user/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要查询的user_id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员根据user_id查询用户完整信息
      tags:
      - 用户相关接口
  /user/get/login:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取当前登录用户
      tags:
      - 用户相关接口
  /user/get/vo:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要查找的用户id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 根据user_id查找用户视图
      tags:
      - 用户相关接口
  /user/list/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页查询所需要的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员根据查询用户完整信息列表
      tags:
      - 用户相关接口
  /user/list/page/vo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分页查询需要的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户视图列表
      tags:
      - 用户相关接口
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.UserDTOLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 用户相关接口
  /user/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户登出
      tags:
      - 用户相关接口
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.UserDTORegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 用户注册
      tags:
      - 用户相关接口
  /user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生，需为管理员
        in: header
        name: Authorization
        required: true
        type: string
      - description: 需要更新的用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.UserDTOUpdateByAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员根据user_id更新用户信息
      tags:
      - 用户相关接口
  /user/update/my:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌 Token 登录后产生
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改后的数据
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.UserDTOUpdateBySelf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 当前用户更新自己的信息
      tags:
      - 用户相关接口
swagger: "2.0"